+++ src/main/java/org/cicirello/search/evo/GenerationalEvolutionaryAlgorithm.java
@@ -21,8 +21,6 @@
 package org.cicirello.search.evo;
 
 import org.cicirello.util.Copyable;
-import org.cicirello.search.ReoptimizableMetaheuristic;
-import org.cicirello.search.SolutionCostPair;
 import org.cicirello.search.ProgressTracker;
 import org.cicirello.search.problems.Problem;
 import org.cicirello.search.operators.Initializer;
@@ -43,7 +41,7 @@ import org.cicirello.math.rand.RandomVariates;
  * {@link MutationOperator}, and {@link SelectionOperator} classes to one of the
  * constructors.</p>
  *
- * <p>This class supports a variety of evolutionary algorithm models, depending
+ * <p>This class supports multiple evolutionary algorithm models, depending
  * upon the constructor you use, including:</p>
  * <ul>
  * <li>The typical generational model using both crossover and mutation, controlled by
@@ -54,22 +52,17 @@ import org.cicirello.math.rand.RandomVariates;
  * controlled by a crossover rate and a mutation rate, but such that each child is the
  * result of crossover, or mutation, or a simply copy of a parent, but never the result
  * of both crossover and mutation.</li>
- * <li>A generational mutation-only evolutionary algorithm.</li>
  * </ul>
- * <p>Note that it does not include a constructor dedicated to a crossover-only case
- * since it would be rare (if ever) that you would find it desirable not to use a 
- * mutation operator. However, if you find a crossover-only use-case, then simply
- * pass any mutation operator and 0.0 for the mutation rate to one of the constructors.</p>
+ * <p>The library also includes a class for mutation-only generational EAs
+ * (see {@link GenerationalMutationOnlyEvolutionaryAlgorithm}).</p>
  *
  * @param <T> The type of object under optimization.
  *
  * @author <a href=https://www.cicirello.org/ target=_top>Vincent A. Cicirello</a>, 
  * <a href=https://www.cicirello.org/ target=_top>https://www.cicirello.org/</a>
  */
-public class GenerationalEvolutionaryAlgorithm<T extends Copyable<T>> implements ReoptimizableMetaheuristic<T> {
+public class GenerationalEvolutionaryAlgorithm<T extends Copyable<T>> extends AbstractEvolutionaryAlgorithm<T> {
 	
-	private final Population<T> pop;
-	private final Problem<T> problem;
 	private final MutationOperator<T> mutation;
 	private final double M;
 	private final CrossoverOperator<T> crossover;
@@ -77,8 +70,6 @@ public class GenerationalEvolutionaryAlgorithm<T extends Copyable<T>> implements
 	
 	private final SingleGen<T> sr;
 	private final GenerationOption go;
-	
-	private long numFitnessEvals;
 	
 	// Constructors for standard generational model using both crossover and mutation operators
 	
@@ -328,110 +319,13 @@ public class GenerationalEvolutionaryAlgorithm<T extends Copyable<T>> implements
 		this(new BasePopulation.Integer<T>(n, initializer, f, selection, new ProgressTracker<T>()), f.getProblem(), mutation, mutationRate, crossover, crossoverRate, mutuallyExclusiveOps);
 	}
 	
-	
-	// Mutation-Only Constructors
-	
-	/**
-	 * Constructs and initializes the evolutionary algorithm with mutation only. This constructor supports fitness functions
-	 * with fitnesses of type double, the {@link FitnessFunction.Double} interface.
-	 *
-	 * @param n The population size.
-	 * @param mutation The mutation operator.
-	 * @param mutationRate The probability that a member of the population is mutated once during a generation. Note that
-	 *     this is not a per-bit rate since this class is generalized to evolution of any {@link Copyable} object type.
-	 *     For {@link org.cicirello.search.representations.BitVector} optimization and traditional genetic algorithm 
-	 *     interpretation of mutation rate, configure
-	 *     your mutation operator with the per-bit mutation rate, and then pass 1.0 for this parameter.
-	 * @param initializer An initializer for generating random initial population members.
-	 * @param f The fitness function.
-	 * @param selection The selection operator.
-	 * @param tracker A ProgressTracker.
-	 *
-	 * @throws IllegalArgumentException if n is less than 1.
-	 * @throws IllegalArgumentException if mutationRate is less than 0.
-	 * @throws NullPointerException if any of mutation, initializer, f, selection, or tracker are null.
-	 */
-	public GenerationalEvolutionaryAlgorithm(int n, MutationOperator<T> mutation, double mutationRate, Initializer<T> initializer, FitnessFunction.Double<T> f, SelectionOperator selection, ProgressTracker<T> tracker) {
-		this(new BasePopulation.Double<T>(n, initializer, f, selection, tracker), f.getProblem(), mutation, mutationRate);
-	}
-	
-	/**
-	 * Constructs and initializes the evolutionary algorithm with mutation only. This constructor supports fitness functions
-	 * with fitnesses of type int, the {@link FitnessFunction.Integer} interface.
-	 *
-	 * @param n The population size.
-	 * @param mutation The mutation operator.
-	 * @param mutationRate The probability that a member of the population is mutated once during a generation. Note that
-	 *     this is not a per-bit rate since this class is generalized to evolution of any {@link Copyable} object type.
-	 *     For {@link org.cicirello.search.representations.BitVector} optimization and traditional genetic algorithm 
-	 *     interpretation of mutation rate, configure
-	 *     your mutation operator with the per-bit mutation rate, and then pass 1.0 for this parameter.
-	 * @param initializer An initializer for generating random initial population members.
-	 * @param f The fitness function.
-	 * @param selection The selection operator.
-	 * @param tracker A ProgressTracker.
-	 *
-	 * @throws IllegalArgumentException if n is less than 1.
-	 * @throws IllegalArgumentException if mutationRate is less than 0.
-	 * @throws NullPointerException if any of mutation, initializer, f, selection, or tracker are null.
-	 */
-	public GenerationalEvolutionaryAlgorithm(int n, MutationOperator<T> mutation, double mutationRate, Initializer<T> initializer, FitnessFunction.Integer<T> f, SelectionOperator selection, ProgressTracker<T> tracker) {
-		this(new BasePopulation.Integer<T>(n, initializer, f, selection, tracker), f.getProblem(), mutation, mutationRate);
-	}
-	
-	/**
-	 * Constructs and initializes the evolutionary algorithm with mutation only. This constructor supports fitness functions
-	 * with fitnesses of type double, the {@link FitnessFunction.Double} interface.
-	 *
-	 * @param n The population size.
-	 * @param mutation The mutation operator.
-	 * @param mutationRate The probability that a member of the population is mutated once during a generation. Note that
-	 *     this is not a per-bit rate since this class is generalized to evolution of any {@link Copyable} object type.
-	 *     For {@link org.cicirello.search.representations.BitVector} optimization and traditional genetic algorithm 
-	 *     interpretation of mutation rate, configure
-	 *     your mutation operator with the per-bit mutation rate, and then pass 1.0 for this parameter.
-	 * @param initializer An initializer for generating random initial population members.
-	 * @param f The fitness function.
-	 * @param selection The selection operator.
-	 *
-	 * @throws IllegalArgumentException if n is less than 1.
-	 * @throws IllegalArgumentException if mutationRate is less than 0.
-	 * @throws NullPointerException if any of mutation, initializer, f, or selection are null.
-	 */
-	public GenerationalEvolutionaryAlgorithm(int n, MutationOperator<T> mutation, double mutationRate, Initializer<T> initializer, FitnessFunction.Double<T> f, SelectionOperator selection) {
-		this(new BasePopulation.Double<T>(n, initializer, f, selection, new ProgressTracker<T>()), f.getProblem(), mutation, mutationRate);
-	}
-	
-	/**
-	 * Constructs and initializes the evolutionary algorithm with mutation only. This constructor supports fitness functions
-	 * with fitnesses of type int, the {@link FitnessFunction.Integer} interface.
-	 *
-	 * @param n The population size.
-	 * @param mutation The mutation operator.
-	 * @param mutationRate The probability that a member of the population is mutated once during a generation. Note that
-	 *     this is not a per-bit rate since this class is generalized to evolution of any {@link Copyable} object type.
-	 *     For {@link org.cicirello.search.representations.BitVector} optimization and traditional genetic algorithm 
-	 *     interpretation of mutation rate, configure
-	 *     your mutation operator with the per-bit mutation rate, and then pass 1.0 for this parameter.
-	 * @param initializer An initializer for generating random initial population members.
-	 * @param f The fitness function.
-	 * @param selection The selection operator.
-	 *
-	 * @throws IllegalArgumentException if n is less than 1.
-	 * @throws IllegalArgumentException if mutationRate is less than 0.
-	 * @throws NullPointerException if any of mutation, initializer, f, or selection are null.
-	 */
-	public GenerationalEvolutionaryAlgorithm(int n, MutationOperator<T> mutation, double mutationRate, Initializer<T> initializer, FitnessFunction.Integer<T> f, SelectionOperator selection) {
-		this(new BasePopulation.Integer<T>(n, initializer, f, selection, new ProgressTracker<T>()), f.getProblem(), mutation, mutationRate);
-	}
-	
-	
 	// Internal Constructors
 	
 	/*
 	 * Internal helper constructor for standard EAs with full generation (both crossover and mutation).
 	 */
 	private GenerationalEvolutionaryAlgorithm(Population<T> pop, Problem<T> problem, MutationOperator<T> mutation, double mutationRate, CrossoverOperator<T> crossover, double crossoverRate, boolean mutuallyExclusiveOps) {
+		super(pop, problem);
 		if (mutation == null) {
 			throw new NullPointerException("mutation must be non-null");
 		}
@@ -462,37 +356,8 @@ public class GenerationalEvolutionaryAlgorithm<T extends Copyable<T>> implements
 			go = GenerationOption.FULL_GENERATION_ALWAYS_MUTATE;
 		}
 		C = crossoverRate < 1.0 ? crossoverRate : 1.0;
-		this.pop = pop;
-		this.problem = problem;
 		this.mutation = mutation;
 		this.crossover = crossover;
-	}
-	
-	/*
-	 * Internal helper constructor for Mutation-Only EAs.
-	 */
-	private GenerationalEvolutionaryAlgorithm(Population<T> pop, Problem<T> problem, MutationOperator<T> mutation, double mutationRate) {
-		if (mutation == null) {
-			throw new NullPointerException("mutation must be non-null");
-		}
-		if (mutationRate < 0.0) {
-			throw new IllegalArgumentException("mutationRate must not be negative");
-		}
-		this.pop = pop;
-		this.problem = problem;
-		this.mutation = mutation;
-		crossover = null;
-		C = 0.0;
-		
-		if (mutationRate < 1.0) {
-			M = mutationRate;
-			sr = mutationOnly(); 
-			go = GenerationOption.MUTATION_ONLY;
-		} else {
-			M = 1.0;
-			sr = alwaysMutate();
-			go = GenerationOption.ALWAYS_MUTATION;
-		}
 	}
 	
 	/*
@@ -500,122 +365,46 @@ public class GenerationalEvolutionaryAlgorithm<T extends Copyable<T>> implements
 	 * package private so subclasses in same package can use it for initialization for their own split methods.
 	 */
 	GenerationalEvolutionaryAlgorithm(GenerationalEvolutionaryAlgorithm<T> other) {
+		super(other);
+		
 		// Must be split
-		pop = other.pop.split();
 		mutation = other.mutation.split();
-		crossover = other.crossover != null ? other.crossover.split() : null; 
+		crossover = other.crossover.split(); 
 		
 		// Threadsafe so just copy reference or values
-		problem = other.problem;
 		M = other.M;
 		C = other.C;
-		
-		// Each instance must maintain its own count of evals.
-		numFitnessEvals = 0;
 		
 		// Initialize the runner   
 		go = other.go;
 		switch (go) {
 			case FULL_GENERATION: sr = fullGeneration(); break;
 			case FULL_GENERATION_ALWAYS_MUTATE: sr = alwaysMutateFullGeneration(); break;
-			case MUTUALLY_EXCLUSIVE_OPERATORS: sr = mutuallyExclusiveOperators(); break;
-			case MUTATION_ONLY: sr = mutationOnly(); break;
-			default: //case ALWAYS_MUTATION: 
-				sr = alwaysMutate(); break;
+			default: //case MUTUALLY_EXCLUSIVE_OPERATORS: 
+				sr = mutuallyExclusiveOperators(); break;
 		}
-	}
-	
-	/**
-	 * Runs the evolutionary algorithm beginning from a randomly generated population. If this 
-	 * method is called multiple times, each call begins at a new randomly generated population.
-	 *
-	 * @param numGenerations The number of generations to run.
-	 *
-	 * @return The best solution found during this set of generations, which may or may not be the
-	 * same as the solution contained in the {@link ProgressTracker}, which contains the best across all
-	 * calls to optimize as well as {@link #reoptimize}. Returns null if the run did not execute, such 
-	 * as if the ProgressTracker already contains the theoretical best solution.
-	 */
-	@Override
-	public final SolutionCostPair<T> optimize(int numGenerations) {
-		if (pop.evolutionIsPaused()) return null;
-		pop.init();
-		pop.initOperators(numGenerations);
-		numFitnessEvals = numFitnessEvals + pop.size();
-		internalOptimize(numGenerations);
-		return pop.getMostFit();
-	}
-	
-	/**
-	 * Runs the evolutionary algorithm continuing from the final population from the most recent call
-	 * to either {@link #optimize} or {@link #reoptimize}, or from a random population if this is the first
-	 * call to either method. 
-	 *
-	 * @param numGenerations The number of generations to run.
-	 *
-	 * @return The best solution found during this set of generations, which may or may not be the
-	 * same as the solution contained in the {@link ProgressTracker}, which contains the best across all
-	 * calls to optimize as well as {@link #optimize}. Returns null if the run did not execute, such 
-	 * as if the ProgressTracker already contains the theoretical best solution.
-	 */
-	@Override
-	public final SolutionCostPair<T> reoptimize(int numGenerations) {
-		if (pop.evolutionIsPaused()) return null;
-		pop.initOperators(numGenerations);
-		internalOptimize(numGenerations);
-		return pop.getMostFit();
 	}
 	
 	@Override
 	public GenerationalEvolutionaryAlgorithm<T> split() {
 		return new GenerationalEvolutionaryAlgorithm<T>(this);
 	}
-	
-	@Override
-	public final Problem<T> getProblem() {
-		return problem;
-	}
-	
-	@Override
-	public final ProgressTracker<T> getProgressTracker() {
-		return pop.getProgressTracker();
-	}
-	
-	@Override
-	public final void setProgressTracker(ProgressTracker<T> tracker) {
-		pop.setProgressTracker(tracker);
-	}
-	
-	/**
-	 * Gets the total run length in number of fitness evaluations. This is the total run length across all 
-	 * calls to {@link #optimize} and {@link #reoptimize}. This may differ from what may be expected 
-	 * based on run lengths. For example, the search terminates if it finds the theoretical best 
-	 * solution, and also immediately returns if a prior call found the theoretical best. In such 
-	 * cases, the total run length may be less than the requested run length.
-	 *
-	 * @return The total number of generations completed across all calls to {@link #optimize} and {@link #reoptimize}.
-	 */
-	@Override
-	public long getTotalRunLength() {
-		return numFitnessEvals;
-	}
 		
 	private interface SingleGen<T extends Copyable<T>> {
-		void optimizeSingleGen();
+		int optimizeSingleGen(Population<T> pop);
 	}
 	
 	private enum GenerationOption {
-		FULL_GENERATION, FULL_GENERATION_ALWAYS_MUTATE, MUTUALLY_EXCLUSIVE_OPERATORS, MUTATION_ONLY, ALWAYS_MUTATION
+		FULL_GENERATION, FULL_GENERATION_ALWAYS_MUTATE, MUTUALLY_EXCLUSIVE_OPERATORS
 	}
 	
-	private void internalOptimize(int numGenerations) {
-		for (int i = 0; i < numGenerations && !pop.evolutionIsPaused(); i++) {
-			sr.optimizeSingleGen();
-		}
+	@Override
+	final int oneGeneration(Population<T> pop) {
+		return sr.optimizeSingleGen(pop);
 	}
 	
 	private SingleGen<T> mutuallyExclusiveOperators() {
-		return () -> {
+		return (pop) -> {
 			pop.select();
 			// Since select() above randomizes ordering, just use a binomial
 			// to get count of number of pairs of parents to cross and cross the first 
@@ -640,12 +429,12 @@ public class GenerationalEvolutionaryAlgorithm<T extends Copyable<T>> implements
 				}
 			}
 			pop.replace();
-			numFitnessEvals = numFitnessEvals + crossed + mutateCount;
+			return crossed + mutateCount;
 		};
 	}
 	
 	private SingleGen<T> fullGeneration() {
-		return () -> {
+		return (pop) -> {
 			pop.select();
 			// Since select() above randomizes ordering, just use a binomial
 			// to get count of number of pairs of parents to cross and cross the first 
@@ -663,12 +452,12 @@ public class GenerationalEvolutionaryAlgorithm<T extends Copyable<T>> implements
 				pop.updateFitness(operateOnThese[j]);
 			}
 			pop.replace();
-			numFitnessEvals = numFitnessEvals + operateOnThese.length + count + count;
+			return operateOnThese.length + count + count;
 		};
 	}
 	
 	private SingleGen<T> alwaysMutateFullGeneration() {
-		return () -> {
+		return (pop) -> {
 			pop.select();
 			// Since select() above randomizes ordering, just use a binomial
 			// to get count of number of pairs of parents to cross and cross the first 
@@ -686,36 +475,7 @@ public class GenerationalEvolutionaryAlgorithm<T extends Copyable<T>> implements
 				pop.updateFitness(j);
 			}
 			pop.replace();
-			numFitnessEvals = numFitnessEvals + LAMBDA + count + count;
+			return LAMBDA + count + count;
 		};
 	}
-	
-	private SingleGen<T> mutationOnly() {
-		return () -> {
-			pop.select();
-			// Since select() above randomizes ordering, just use a binomial
-			// to get count of how many to mutate and mutate the first count individuals.
-			final int count = RandomVariates.nextBinomial(pop.mutableSize(), M);
-			for (int j = 0; j < count; j++) {
-				mutation.mutate(pop.get(j));
-				pop.updateFitness(j);
-			}
-			pop.replace();
-			numFitnessEvals = numFitnessEvals + count;
-		};
-	}
-	
-	private SingleGen<T> alwaysMutate() {
-		return () -> {
-			final int LAMBDA = pop.mutableSize();
-			pop.select();
-			for (int j = 0; j < LAMBDA; j++) {
-				mutation.mutate(pop.get(j));
-				pop.updateFitness(j);
-			}
-			pop.replace();
-			numFitnessEvals = numFitnessEvals + LAMBDA;
-		};
-	}
-	
 }