+++ src/test/java/org/cicirello/search/evo/GenerationalEANoPTTests.java
@@ -33,7 +33,9 @@ import org.cicirello.search.operators.MutationOperator;
 import org.cicirello.search.operators.CrossoverOperator;
 
 /**
- * JUnit 4 test cases for GenerationalEvolutionaryAlgorithm, cases with constructors without ProgressTracker parameter.
+ * JUnit 4 test cases for GenerationalEvolutionaryAlgorithm and
+ * GenerationalMutationOnlyEvolutionaryAlgorithm, cases with constructors 
+ * without ProgressTracker parameter.
  */
 public class GenerationalEANoPTTests {
 	
@@ -41,7 +43,7 @@ public class GenerationalEANoPTTests {
 	public void testExceptions() {
 		NullPointerException thrownNull = assertThrows( 
 			NullPointerException.class,
-			() -> new GenerationalEvolutionaryAlgorithm<TestObject>(
+			() -> new GenerationalMutationOnlyEvolutionaryAlgorithm<TestObject>(
 				5, 
 				null, 
 				0.5, 
@@ -52,7 +54,7 @@ public class GenerationalEANoPTTests {
 		);
 		thrownNull = assertThrows( 
 			NullPointerException.class,
-			() -> new GenerationalEvolutionaryAlgorithm<TestObject>(
+			() -> new GenerationalMutationOnlyEvolutionaryAlgorithm<TestObject>(
 				5, 
 				null, 
 				0.5, 
@@ -171,7 +173,7 @@ public class GenerationalEANoPTTests {
 		);
 		IllegalArgumentException thrownIllegal = assertThrows( 
 			IllegalArgumentException.class,
-			() -> new GenerationalEvolutionaryAlgorithm<TestObject>(
+			() -> new GenerationalMutationOnlyEvolutionaryAlgorithm<TestObject>(
 				5, 
 				new CountMutationCalls(), 
 				-1E-10, 
@@ -182,7 +184,7 @@ public class GenerationalEANoPTTests {
 		);
 		thrownIllegal = assertThrows( 
 			IllegalArgumentException.class,
-			() -> new GenerationalEvolutionaryAlgorithm<TestObject>(
+			() -> new GenerationalMutationOnlyEvolutionaryAlgorithm<TestObject>(
 				5, 
 				new CountMutationCalls(), 
 				-1E-10, 
@@ -263,9 +265,9 @@ public class GenerationalEANoPTTests {
 	public void testStoppedByTracker() {
 		class TestThread extends Thread {
 			
-			GenerationalEvolutionaryAlgorithm<TestObject> ea;
+			GenerationalMutationOnlyEvolutionaryAlgorithm<TestObject> ea;
 			
-			public TestThread(GenerationalEvolutionaryAlgorithm<TestObject> ea) {
+			public TestThread(GenerationalMutationOnlyEvolutionaryAlgorithm<TestObject> ea) {
 				this.ea = ea;
 			}
 			
@@ -282,7 +284,7 @@ public class GenerationalEANoPTTests {
 		CountMutationCalls mutation = new CountMutationCalls();
 		final int N = 100;
 		
-		GenerationalEvolutionaryAlgorithm<TestObject> ea = new GenerationalEvolutionaryAlgorithm<TestObject>(
+		GenerationalMutationOnlyEvolutionaryAlgorithm<TestObject> ea = new GenerationalMutationOnlyEvolutionaryAlgorithm<TestObject>(
 			N, 
 			mutation, 
 			0.5, 
@@ -597,7 +599,7 @@ public class GenerationalEANoPTTests {
 		CountMutationCalls mutation = new CountMutationCalls();
 		final int N = 100;
 		
-		GenerationalEvolutionaryAlgorithm<TestObject> ea = new GenerationalEvolutionaryAlgorithm<TestObject>(
+		GenerationalMutationOnlyEvolutionaryAlgorithm<TestObject> ea = new GenerationalMutationOnlyEvolutionaryAlgorithm<TestObject>(
 			N, 
 			mutation, 
 			0.5, 
@@ -626,7 +628,7 @@ public class GenerationalEANoPTTests {
 		assertEquals(1, selection.initCount);
 		
 		// split it
-		GenerationalEvolutionaryAlgorithm<TestObject> ea2 = ea.split();
+		GenerationalMutationOnlyEvolutionaryAlgorithm<TestObject> ea2 = ea.split();
 		
 		// reoptimize
 		int oldMutationCount = mutation.count;
@@ -683,7 +685,7 @@ public class GenerationalEANoPTTests {
 		CountMutationCalls mutation = new CountMutationCalls();
 		final int N = 100;
 		
-		GenerationalEvolutionaryAlgorithm<TestObject> ea = new GenerationalEvolutionaryAlgorithm<TestObject>(
+		GenerationalMutationOnlyEvolutionaryAlgorithm<TestObject> ea = new GenerationalMutationOnlyEvolutionaryAlgorithm<TestObject>(
 			N, 
 			mutation, 
 			1.0, 
@@ -712,7 +714,7 @@ public class GenerationalEANoPTTests {
 		assertEquals(1, selection.initCount);
 		
 		// split it
-		GenerationalEvolutionaryAlgorithm<TestObject> ea2 = ea.split();
+		GenerationalMutationOnlyEvolutionaryAlgorithm<TestObject> ea2 = ea.split();
 		
 		// reoptimize
 		solution = ea.reoptimize(5);
@@ -1727,7 +1729,7 @@ public class GenerationalEANoPTTests {
 		CountMutationCalls mutation = new CountMutationCalls();
 		final int N = 100;
 		
-		GenerationalEvolutionaryAlgorithm<TestObject> ea = new GenerationalEvolutionaryAlgorithm<TestObject>(
+		GenerationalMutationOnlyEvolutionaryAlgorithm<TestObject> ea = new GenerationalMutationOnlyEvolutionaryAlgorithm<TestObject>(
 			N, 
 			mutation, 
 			0.5, 
@@ -1755,7 +1757,7 @@ public class GenerationalEANoPTTests {
 		assertEquals(1, selection.initCount);
 		
 		// split it
-		GenerationalEvolutionaryAlgorithm<TestObject> ea2 = ea.split();
+		GenerationalMutationOnlyEvolutionaryAlgorithm<TestObject> ea2 = ea.split();
 		
 		// reoptimize
 		int oldMutationCount = mutation.count;
@@ -1812,7 +1814,7 @@ public class GenerationalEANoPTTests {
 		CountMutationCalls mutation = new CountMutationCalls();
 		final int N = 100;
 		
-		GenerationalEvolutionaryAlgorithm<TestObject> ea = new GenerationalEvolutionaryAlgorithm<TestObject>(
+		GenerationalMutationOnlyEvolutionaryAlgorithm<TestObject> ea = new GenerationalMutationOnlyEvolutionaryAlgorithm<TestObject>(
 			N, 
 			mutation, 
 			1.0, 
@@ -1841,7 +1843,7 @@ public class GenerationalEANoPTTests {
 		assertEquals(1, selection.initCount);
 		
 		// split it
-		GenerationalEvolutionaryAlgorithm<TestObject> ea2 = ea.split();
+		GenerationalMutationOnlyEvolutionaryAlgorithm<TestObject> ea2 = ea.split();
 		
 		// reoptimize
 		solution = ea.reoptimize(5);